:recycle: refactor: refactor the plugin and chat service and plugins selectors entry
