+++ src/store/plugin/index.ts
@@ -1,3 +1,2 @@
 export * from './helpers';
-export * from './selectors';
 export { usePluginStore } from './store';
